refactor(agent-gui): extract visible-error card below the 800-line limit
Keep AgentMessageBlock under the agent-gui degradation budget by moving the
visible-error card and details disclosure into focused components.
