+++ packages/agent/gui/shared/agentEnv/agentErrorPresentation.spec.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
 import {
   classifyFailedAgentMessage,
+  isProviderPlanLimitMessage,
   resolveAgentErrorPresentation
 } from "./agentErrorPresentation";
 
@@ -39,6 +40,19 @@ describe("classifyFailedAgentMessage", () => {
     expect(classifyFailedAgentMessage("Add a payment method to continue")).toBe(
       "quota_or_rate_limit"
     );
+  });
+});
+
+describe("isProviderPlanLimitMessage", () => {
+  it("matches Cursor plan/payment gate copy only", () => {
+    expect(isProviderPlanLimitMessage("Upgrade your plan to continue")).toBe(
+      true
+    );
+    expect(isProviderPlanLimitMessage("Add a payment method to continue")).toBe(
+      true
+    );
+    expect(isProviderPlanLimitMessage("rate limit exceeded")).toBe(false);
+    expect(isProviderPlanLimitMessage("")).toBe(false);
   });
 });
